--- crates/verona-rt/tests/cown-write-stress.rs
@@ -1,609 +1,0 @@
-use verona_rt::cown::CownPtr;
-
-fn stress_once<const N: usize>() {
-    let mut x = CownPtr::new([0u8; N]);
-
-    unsafe {
-        for i in x.yolo_data() {
-            assert_eq!(*i, 0);
-            *i = 0xCC;
-        }
-    }
-
-    let mut x2 = x.clone();
-    unsafe {
-        for i in x2.yolo_data() {
-            assert_eq!(*i, 0xCC);
-            *i = 0x33;
-        }
-    }
-    drop(x2);
-
-    unsafe {
-        for i in x.yolo_data() {
-            assert_eq!(*i, 0x33);
-        }
-    }
-
-    drop(x);
-}
-
-fn repeat_alloc<const N: usize>() {
-    for _ in 0..100 {
-        stress_once::<N>();
-    }
-}
-
-#[test]
-fn main() {
-    unsafe {
-        verona_rt_sys::enable_logging();
-    }
-
-    verona_rt::scheduler::with_leak_detector(|| {
-        repeat_alloc::<4096>();
-        repeat_alloc::<4039>();
-        repeat_alloc::<4038>();
-        repeat_alloc::<4037>();
-        repeat_alloc::<4036>();
-        repeat_alloc::<4035>();
-        repeat_alloc::<4034>();
-        repeat_alloc::<4033>();
-        repeat_alloc::<4032>();
-
-        repeat_alloc::<3932>();
-
-        repeat_alloc::<3719>();
-        repeat_alloc::<3718>();
-
-        repeat_alloc::<1478>();
-        repeat_alloc::<1477>();
-
-        repeat_alloc::<545>();
-        repeat_alloc::<544>();
-        repeat_alloc::<543>();
-        repeat_alloc::<542>();
-        repeat_alloc::<541>();
-        repeat_alloc::<540>();
-        repeat_alloc::<539>();
-        repeat_alloc::<538>();
-        repeat_alloc::<537>();
-        repeat_alloc::<536>();
-        repeat_alloc::<535>();
-        repeat_alloc::<534>();
-        repeat_alloc::<533>();
-        repeat_alloc::<532>();
-        repeat_alloc::<531>();
-        repeat_alloc::<530>();
-        repeat_alloc::<529>();
-        repeat_alloc::<528>();
-        repeat_alloc::<527>();
-        repeat_alloc::<526>();
-        repeat_alloc::<525>();
-        repeat_alloc::<524>();
-        repeat_alloc::<523>();
-        repeat_alloc::<522>();
-        repeat_alloc::<521>();
-        repeat_alloc::<520>();
-        repeat_alloc::<519>();
-        repeat_alloc::<518>();
-        repeat_alloc::<517>();
-        repeat_alloc::<516>();
-        repeat_alloc::<515>();
-        repeat_alloc::<514>();
-        repeat_alloc::<513>();
-        repeat_alloc::<512>();
-        repeat_alloc::<511>();
-        repeat_alloc::<510>();
-        repeat_alloc::<509>();
-        repeat_alloc::<508>();
-        repeat_alloc::<507>();
-        repeat_alloc::<506>();
-        repeat_alloc::<505>();
-        repeat_alloc::<504>();
-        repeat_alloc::<503>();
-        repeat_alloc::<502>();
-        repeat_alloc::<501>();
-        repeat_alloc::<500>();
-        repeat_alloc::<499>();
-        repeat_alloc::<498>();
-        repeat_alloc::<497>();
-        repeat_alloc::<496>();
-        repeat_alloc::<495>();
-        repeat_alloc::<494>();
-        repeat_alloc::<493>();
-        repeat_alloc::<492>();
-        repeat_alloc::<491>();
-        repeat_alloc::<490>();
-        repeat_alloc::<489>();
-        repeat_alloc::<488>();
-        repeat_alloc::<487>();
-        repeat_alloc::<486>();
-        repeat_alloc::<485>();
-        repeat_alloc::<484>();
-        repeat_alloc::<483>();
-        repeat_alloc::<482>();
-        repeat_alloc::<481>();
-        repeat_alloc::<480>();
-        repeat_alloc::<479>();
-        repeat_alloc::<478>();
-        repeat_alloc::<477>();
-        repeat_alloc::<476>();
-        repeat_alloc::<475>();
-        repeat_alloc::<474>();
-        repeat_alloc::<473>();
-        repeat_alloc::<472>();
-        repeat_alloc::<471>();
-        repeat_alloc::<470>();
-        repeat_alloc::<469>();
-        repeat_alloc::<468>();
-        repeat_alloc::<467>();
-        repeat_alloc::<466>();
-        repeat_alloc::<465>();
-        repeat_alloc::<464>();
-        repeat_alloc::<463>();
-        repeat_alloc::<462>();
-        repeat_alloc::<461>();
-        repeat_alloc::<460>();
-        repeat_alloc::<459>();
-        repeat_alloc::<458>();
-        repeat_alloc::<457>();
-        repeat_alloc::<456>();
-        repeat_alloc::<455>();
-        repeat_alloc::<454>();
-        repeat_alloc::<453>();
-        repeat_alloc::<452>();
-        repeat_alloc::<451>();
-        repeat_alloc::<450>();
-        repeat_alloc::<449>();
-        repeat_alloc::<448>();
-        repeat_alloc::<447>();
-        repeat_alloc::<446>();
-        repeat_alloc::<445>();
-        repeat_alloc::<444>();
-        repeat_alloc::<443>();
-        repeat_alloc::<442>();
-        repeat_alloc::<441>();
-        repeat_alloc::<440>();
-        repeat_alloc::<439>();
-        repeat_alloc::<438>();
-        repeat_alloc::<437>();
-        repeat_alloc::<436>();
-        repeat_alloc::<435>();
-        repeat_alloc::<434>();
-        repeat_alloc::<433>();
-        repeat_alloc::<432>();
-        repeat_alloc::<431>();
-        repeat_alloc::<430>();
-        repeat_alloc::<429>();
-        repeat_alloc::<428>();
-        repeat_alloc::<427>();
-        repeat_alloc::<426>();
-        repeat_alloc::<425>();
-        repeat_alloc::<424>();
-        repeat_alloc::<423>();
-        repeat_alloc::<422>();
-        repeat_alloc::<421>();
-        repeat_alloc::<420>();
-        repeat_alloc::<419>();
-        repeat_alloc::<418>();
-        repeat_alloc::<417>();
-        repeat_alloc::<416>();
-        repeat_alloc::<415>();
-        repeat_alloc::<414>();
-        repeat_alloc::<413>();
-        repeat_alloc::<412>();
-        repeat_alloc::<411>();
-        repeat_alloc::<410>();
-        repeat_alloc::<409>();
-        repeat_alloc::<408>();
-        repeat_alloc::<407>();
-        repeat_alloc::<406>();
-        repeat_alloc::<405>();
-        repeat_alloc::<404>();
-        repeat_alloc::<403>();
-        repeat_alloc::<402>();
-        repeat_alloc::<401>();
-        repeat_alloc::<400>();
-        repeat_alloc::<399>();
-        repeat_alloc::<398>();
-        repeat_alloc::<397>();
-        repeat_alloc::<396>();
-        repeat_alloc::<395>();
-        repeat_alloc::<394>();
-        repeat_alloc::<393>();
-        repeat_alloc::<392>();
-        repeat_alloc::<391>();
-        repeat_alloc::<390>();
-        repeat_alloc::<389>();
-        repeat_alloc::<388>();
-        repeat_alloc::<387>();
-        repeat_alloc::<386>();
-        repeat_alloc::<385>();
-        repeat_alloc::<384>();
-        repeat_alloc::<383>();
-        repeat_alloc::<382>();
-        repeat_alloc::<381>();
-        repeat_alloc::<380>();
-        repeat_alloc::<379>();
-        repeat_alloc::<378>();
-        repeat_alloc::<377>();
-        repeat_alloc::<376>();
-        repeat_alloc::<375>();
-        repeat_alloc::<374>();
-        repeat_alloc::<373>();
-        repeat_alloc::<372>();
-        repeat_alloc::<371>();
-        repeat_alloc::<370>();
-        repeat_alloc::<369>();
-        repeat_alloc::<368>();
-        repeat_alloc::<367>();
-        repeat_alloc::<366>();
-        repeat_alloc::<365>();
-        repeat_alloc::<364>();
-        repeat_alloc::<363>();
-        repeat_alloc::<362>();
-        repeat_alloc::<361>();
-        repeat_alloc::<360>();
-        repeat_alloc::<359>();
-        repeat_alloc::<358>();
-        repeat_alloc::<357>();
-        repeat_alloc::<356>();
-        repeat_alloc::<355>();
-        repeat_alloc::<354>();
-        repeat_alloc::<353>();
-        repeat_alloc::<352>();
-        repeat_alloc::<351>();
-        repeat_alloc::<350>();
-        repeat_alloc::<349>();
-        repeat_alloc::<348>();
-        repeat_alloc::<347>();
-        repeat_alloc::<346>();
-        repeat_alloc::<345>();
-        repeat_alloc::<344>();
-        repeat_alloc::<343>();
-        repeat_alloc::<342>();
-        repeat_alloc::<341>();
-        repeat_alloc::<340>();
-        repeat_alloc::<339>();
-        repeat_alloc::<338>();
-        repeat_alloc::<337>();
-        repeat_alloc::<336>();
-        repeat_alloc::<335>();
-        repeat_alloc::<334>();
-        repeat_alloc::<333>();
-        repeat_alloc::<332>();
-        repeat_alloc::<331>();
-        repeat_alloc::<330>();
-        repeat_alloc::<329>();
-        repeat_alloc::<328>();
-        repeat_alloc::<327>();
-        repeat_alloc::<326>();
-        repeat_alloc::<325>();
-        repeat_alloc::<324>();
-        repeat_alloc::<323>();
-        repeat_alloc::<322>();
-        repeat_alloc::<321>();
-        repeat_alloc::<320>();
-        repeat_alloc::<319>();
-        repeat_alloc::<318>();
-        repeat_alloc::<317>();
-        repeat_alloc::<316>();
-        repeat_alloc::<315>();
-        repeat_alloc::<314>();
-        repeat_alloc::<313>();
-        repeat_alloc::<312>();
-        repeat_alloc::<311>();
-        repeat_alloc::<310>();
-        repeat_alloc::<309>();
-        repeat_alloc::<308>();
-        repeat_alloc::<307>();
-        repeat_alloc::<306>();
-        repeat_alloc::<305>();
-        repeat_alloc::<304>();
-        repeat_alloc::<303>();
-        repeat_alloc::<302>();
-        repeat_alloc::<301>();
-        repeat_alloc::<300>();
-        repeat_alloc::<299>();
-        repeat_alloc::<298>();
-        repeat_alloc::<297>();
-        repeat_alloc::<296>();
-        repeat_alloc::<295>();
-        repeat_alloc::<294>();
-        repeat_alloc::<293>();
-        repeat_alloc::<292>();
-        repeat_alloc::<291>();
-        repeat_alloc::<290>();
-        repeat_alloc::<289>();
-        repeat_alloc::<288>();
-        repeat_alloc::<287>();
-        repeat_alloc::<286>();
-        repeat_alloc::<285>();
-        repeat_alloc::<284>();
-        repeat_alloc::<283>();
-        repeat_alloc::<282>();
-        repeat_alloc::<281>();
-        repeat_alloc::<280>();
-        repeat_alloc::<279>();
-        repeat_alloc::<278>();
-        repeat_alloc::<277>();
-        repeat_alloc::<276>();
-        repeat_alloc::<275>();
-        repeat_alloc::<274>();
-        repeat_alloc::<273>();
-        repeat_alloc::<272>();
-        repeat_alloc::<271>();
-        repeat_alloc::<270>();
-        repeat_alloc::<269>();
-        repeat_alloc::<268>();
-        repeat_alloc::<267>();
-        repeat_alloc::<266>();
-        repeat_alloc::<265>();
-        repeat_alloc::<264>();
-        repeat_alloc::<263>();
-        repeat_alloc::<262>();
-        repeat_alloc::<261>();
-        repeat_alloc::<260>();
-        repeat_alloc::<259>();
-        repeat_alloc::<258>();
-        repeat_alloc::<257>();
-        repeat_alloc::<256>();
-        repeat_alloc::<255>();
-        repeat_alloc::<254>();
-        repeat_alloc::<253>();
-        repeat_alloc::<252>();
-        repeat_alloc::<251>();
-        repeat_alloc::<250>();
-        repeat_alloc::<249>();
-        repeat_alloc::<248>();
-        repeat_alloc::<247>();
-        repeat_alloc::<246>();
-        repeat_alloc::<245>();
-        repeat_alloc::<244>();
-        repeat_alloc::<243>();
-        repeat_alloc::<242>();
-        repeat_alloc::<241>();
-        repeat_alloc::<240>();
-        repeat_alloc::<239>();
-        repeat_alloc::<238>();
-        repeat_alloc::<237>();
-        repeat_alloc::<236>();
-        repeat_alloc::<235>();
-        repeat_alloc::<234>();
-        repeat_alloc::<233>();
-        repeat_alloc::<232>();
-        repeat_alloc::<231>();
-        repeat_alloc::<230>();
-        repeat_alloc::<229>();
-        repeat_alloc::<228>();
-        repeat_alloc::<227>();
-        repeat_alloc::<226>();
-        repeat_alloc::<225>();
-        repeat_alloc::<224>();
-        repeat_alloc::<223>();
-        repeat_alloc::<222>();
-        repeat_alloc::<221>();
-        repeat_alloc::<220>();
-        repeat_alloc::<219>();
-        repeat_alloc::<218>();
-        repeat_alloc::<217>();
-        repeat_alloc::<216>();
-        repeat_alloc::<215>();
-        repeat_alloc::<214>();
-        repeat_alloc::<213>();
-        repeat_alloc::<212>();
-        repeat_alloc::<211>();
-        repeat_alloc::<210>();
-        repeat_alloc::<209>();
-        repeat_alloc::<208>();
-        repeat_alloc::<207>();
-        repeat_alloc::<206>();
-        repeat_alloc::<205>();
-        repeat_alloc::<204>();
-        repeat_alloc::<203>();
-        repeat_alloc::<202>();
-        repeat_alloc::<201>();
-        repeat_alloc::<200>();
-        repeat_alloc::<199>();
-        repeat_alloc::<198>();
-        repeat_alloc::<197>();
-        repeat_alloc::<196>();
-        repeat_alloc::<195>();
-        repeat_alloc::<194>();
-        repeat_alloc::<193>();
-        repeat_alloc::<192>();
-        repeat_alloc::<191>();
-        repeat_alloc::<190>();
-        repeat_alloc::<189>();
-        repeat_alloc::<188>();
-        repeat_alloc::<187>();
-        repeat_alloc::<186>();
-        repeat_alloc::<185>();
-        repeat_alloc::<184>();
-        repeat_alloc::<183>();
-        repeat_alloc::<182>();
-        repeat_alloc::<181>();
-        repeat_alloc::<180>();
-        repeat_alloc::<179>();
-        repeat_alloc::<178>();
-        repeat_alloc::<177>();
-        repeat_alloc::<176>();
-        repeat_alloc::<175>();
-        repeat_alloc::<174>();
-        repeat_alloc::<173>();
-        repeat_alloc::<172>();
-        repeat_alloc::<171>();
-        repeat_alloc::<170>();
-        repeat_alloc::<169>();
-        repeat_alloc::<168>();
-        repeat_alloc::<167>();
-        repeat_alloc::<166>();
-        repeat_alloc::<165>();
-        repeat_alloc::<164>();
-        repeat_alloc::<163>();
-        repeat_alloc::<162>();
-        repeat_alloc::<161>();
-        repeat_alloc::<160>();
-        repeat_alloc::<159>();
-        repeat_alloc::<158>();
-        repeat_alloc::<157>();
-        repeat_alloc::<156>();
-        repeat_alloc::<155>();
-        repeat_alloc::<154>();
-        repeat_alloc::<153>();
-        repeat_alloc::<152>();
-        repeat_alloc::<151>();
-        repeat_alloc::<150>();
-        repeat_alloc::<149>();
-        repeat_alloc::<148>();
-        repeat_alloc::<147>();
-        repeat_alloc::<146>();
-        repeat_alloc::<145>();
-        repeat_alloc::<144>();
-        repeat_alloc::<143>();
-        repeat_alloc::<142>();
-        repeat_alloc::<141>();
-        repeat_alloc::<140>();
-        repeat_alloc::<139>();
-        repeat_alloc::<138>();
-        repeat_alloc::<137>();
-        repeat_alloc::<136>();
-        repeat_alloc::<135>();
-        repeat_alloc::<134>();
-        repeat_alloc::<133>();
-        repeat_alloc::<132>();
-        repeat_alloc::<131>();
-        repeat_alloc::<130>();
-        repeat_alloc::<129>();
-        repeat_alloc::<128>();
-        repeat_alloc::<127>();
-        repeat_alloc::<126>();
-        repeat_alloc::<125>();
-        repeat_alloc::<124>();
-        repeat_alloc::<123>();
-        repeat_alloc::<122>();
-        repeat_alloc::<121>();
-        repeat_alloc::<120>();
-        repeat_alloc::<119>();
-        repeat_alloc::<118>();
-        repeat_alloc::<117>();
-        repeat_alloc::<116>();
-        repeat_alloc::<115>();
-        repeat_alloc::<114>();
-        repeat_alloc::<113>();
-        repeat_alloc::<112>();
-        repeat_alloc::<111>();
-        repeat_alloc::<110>();
-        repeat_alloc::<109>();
-        repeat_alloc::<108>();
-        repeat_alloc::<107>();
-        repeat_alloc::<106>();
-        repeat_alloc::<105>();
-        repeat_alloc::<104>();
-        repeat_alloc::<103>();
-        repeat_alloc::<102>();
-        repeat_alloc::<101>();
-        repeat_alloc::<100>();
-        repeat_alloc::<99>();
-        repeat_alloc::<98>();
-        repeat_alloc::<97>();
-        repeat_alloc::<96>();
-        repeat_alloc::<95>();
-        repeat_alloc::<94>();
-        repeat_alloc::<93>();
-        repeat_alloc::<92>();
-        repeat_alloc::<91>();
-        repeat_alloc::<90>();
-        repeat_alloc::<89>();
-        repeat_alloc::<88>();
-        repeat_alloc::<87>();
-        repeat_alloc::<86>();
-        repeat_alloc::<85>();
-        repeat_alloc::<84>();
-        repeat_alloc::<83>();
-        repeat_alloc::<82>();
-        repeat_alloc::<81>();
-        repeat_alloc::<80>();
-        repeat_alloc::<79>();
-        repeat_alloc::<78>();
-        repeat_alloc::<77>();
-        repeat_alloc::<76>();
-        repeat_alloc::<75>();
-        repeat_alloc::<74>();
-        repeat_alloc::<73>();
-        repeat_alloc::<72>();
-        repeat_alloc::<71>();
-        repeat_alloc::<70>();
-        repeat_alloc::<69>();
-        repeat_alloc::<68>();
-        repeat_alloc::<67>();
-        repeat_alloc::<66>();
-        repeat_alloc::<65>();
-        repeat_alloc::<64>();
-        repeat_alloc::<63>();
-        repeat_alloc::<62>();
-        repeat_alloc::<61>();
-        repeat_alloc::<60>();
-        repeat_alloc::<59>();
-        repeat_alloc::<58>();
-        repeat_alloc::<57>();
-        repeat_alloc::<56>();
-        repeat_alloc::<55>();
-        repeat_alloc::<54>();
-        repeat_alloc::<53>();
-        repeat_alloc::<52>();
-        repeat_alloc::<51>();
-        repeat_alloc::<50>();
-        repeat_alloc::<49>();
-        repeat_alloc::<48>();
-        repeat_alloc::<47>();
-        repeat_alloc::<46>();
-        repeat_alloc::<45>();
-        repeat_alloc::<44>();
-        repeat_alloc::<43>();
-        repeat_alloc::<42>();
-        repeat_alloc::<41>();
-        repeat_alloc::<40>();
-        repeat_alloc::<39>();
-        repeat_alloc::<38>();
-        repeat_alloc::<37>();
-        repeat_alloc::<36>();
-        repeat_alloc::<35>();
-        repeat_alloc::<34>();
-        repeat_alloc::<33>();
-        repeat_alloc::<32>();
-        repeat_alloc::<31>();
-        repeat_alloc::<30>();
-        repeat_alloc::<29>();
-        repeat_alloc::<28>();
-        repeat_alloc::<27>();
-        repeat_alloc::<26>();
-        repeat_alloc::<25>();
-        repeat_alloc::<24>();
-        repeat_alloc::<23>();
-        repeat_alloc::<22>();
-        repeat_alloc::<21>();
-        repeat_alloc::<20>();
-        repeat_alloc::<19>();
-        repeat_alloc::<18>();
-        repeat_alloc::<17>();
-        repeat_alloc::<16>();
-        repeat_alloc::<15>();
-        repeat_alloc::<14>();
-        repeat_alloc::<13>();
-        repeat_alloc::<12>();
-        repeat_alloc::<11>();
-        repeat_alloc::<10>();
-        repeat_alloc::<9>();
-        repeat_alloc::<8>();
-        repeat_alloc::<7>();
-        repeat_alloc::<6>();
-        repeat_alloc::<5>();
-        repeat_alloc::<4>();
-        repeat_alloc::<3>();
-        repeat_alloc::<2>();
-        repeat_alloc::<1>();
-        repeat_alloc::<0>();
-    });
-}
